// src/types/yandex/notification-types.ts
export interface OrderStatusUpdatedNotificationDTO {
    /**
     * Тип уведомления:
     *
     * * `PING` — проверочное уведомление.
     * * `ORDER_CREATED` — создан новый заказ.
     * * `ORDER_CANCELLED` — заказ отменен.
     * * `ORDER_STATUS_UPDATED` — статус заказа изменен.
     * * `ORDER_RETURN_CREATED` — создан новый невыкуп или возврат.
     * * `ORDER_CANCELLATION_REQUEST` — создана заявка на отмену заказа (для DBS-магазинов).
     * * `ORDER_RETURN_STATUS_UPDATED` — статус невыкупа или возврата изменен.
     * * `GOODS_FEEDBACK_CREATED` — создан новый отзыв о товаре.
     * * `GOODS_FEEDBACK_COMMENT_CREATED` — создан новый комментарий к отзыву о товаре.
     * * `CHAT_CREATED` — создан новый чат с покупателем.
     * * `CHAT_MESSAGE_SENT` — добавлено новое сообщение в чате.
     * * `CHAT_ARBITRAGE_STARTED` — по обращению покупателя начался спор.
     * * `CHAT_ARBITRAGE_FINISHED` — спор завершен.
     */
    notificationType: NotificationType
    /**
     * Идентификатор заказа.
     * @format int64
     */
    orderId: number
    /**
     * Идентификатор магазина.
     * @format int64
     */
    campaignId: number
    /**
     * Статус заказа:
     *
     * * `PLACING` — оформляется, подготовка к резервированию.
     *
     * * `RESERVED` — зарезервирован, но недооформлен.
     *
     * * `UNPAID` — оформлен, но еще не оплачен (если выбрана оплата при оформлении).
     *
     * * `PROCESSING` — находится в обработке.
     *
     * * `DELIVERY` — передан в службу доставки.
     *
     * * `PICKUP` — доставлен в пункт самовывоза.
     *
     * * `DELIVERED` — получен покупателем.
     *
     * * `CANCELLED` — отменен.
     *
     * * `PENDING` — ожидает обработки со стороны продавца.
     *
     * * `PARTIALLY_RETURNED` — возвращен частично.
     *
     * * `RETURNED` — возвращен полностью.
     *
     * * `UNKNOWN` — неизвестный статус.
     *
     * Также могут возвращаться другие значения. Обрабатывать их не требуется.
     */
    status: OrderStatusType
    /**
     * Этап обработки заказа (если он имеет статус `PROCESSING`) или причина отмены заказа (если он имеет статус `CANCELLED`).
     *
     * * Значения для заказа в статусе `PROCESSING`:
     *
     *     * `STARTED` — заказ подтвержден, его можно начать обрабатывать.
     *
     *     * `READY_TO_SHIP` — заказ собран и готов к отправке.
     *
     * * Значения для заказа в статусе `CANCELLED`:
     *
     *     * `RESERVATION_EXPIRED` — покупатель не завершил оформление зарезервированного заказа в течение 10 минут.
     *
     *     * `USER_NOT_PAID` — покупатель не оплатил заказ (для типа оплаты `PREPAID`) в течение 30 минут.
     *
     *     * `USER_UNREACHABLE` — не удалось связаться с покупателем. Для отмены с этой причиной необходимо выполнить условия:
     *
     *       * не менее 3 звонков с 8 до 21 в часовом поясе покупателя;
     *       * перерыв между первым и третьим звонком не менее 90 минут;
     *       * соединение не короче 5 секунд.
     *
     *       Если хотя бы одно из этих условий не выполнено (кроме случая, когда номер недоступен), отменить заказ не получится. Вернется ответ с кодом ошибки 400.
     *
     *     * `USER_CHANGED_MIND` — покупатель отменил заказ по личным причинам.
     *
     *     * `USER_REFUSED_DELIVERY` — покупателя не устроили условия доставки.
     *
     *     * `USER_REFUSED_PRODUCT` — покупателю не подошел товар.
     *
     *     * `SHOP_FAILED` — магазин не может выполнить заказ.
     *
     *     * `USER_REFUSED_QUALITY` — покупателя не устроило качество товара.
     *
     *     * `REPLACING_ORDER` — покупатель решил заменить товар другим по собственной инициативе.
     *
     *     * `PROCESSING_EXPIRED` — значение более не используется.
     *
     *     * `PICKUP_EXPIRED` — закончился срок хранения заказа в ПВЗ.
     *
     *     * `TOO_MANY_DELIVERY_DATE_CHANGES` — заказ переносили слишком много раз.
     *
     *     * `TOO_LONG_DELIVERY` — заказ доставляется слишком долго.
     *
     *
     * * `TECHNICAL_ERROR` — техническая ошибка на стороне Маркета. Обратитесь в поддержку.
     *
     * Также могут возвращаться другие значения. Обрабатывать их не требуется.
     */
    substatus: OrderSubstatusType
    /**
     * Дата и время изменения статуса заказа.
     *
     * Формат даты: ISO 8601 со смещением относительно UTC. Например, `2017-11-21T00:00:00.213Z`.
     * @format date-time
     */
    updatedAt: string
}

/**
 * Уведомление об отмене заказа.
 *
 * `notificationType` = `ORDER_CANCELLED`
 */
export interface OrderCancelledNotificationDTO {
    /**
     * Тип уведомления:
     *
     * * `PING` — проверочное уведомление.
     * * `ORDER_CREATED` — создан новый заказ.
     * * `ORDER_CANCELLED` — заказ отменен.
     * * `ORDER_STATUS_UPDATED` — статус заказа изменен.
     * * `ORDER_RETURN_CREATED` — создан новый невыкуп или возврат.
     * * `ORDER_CANCELLATION_REQUEST` — создана заявка на отмену заказа (для DBS-магазинов).
     * * `ORDER_RETURN_STATUS_UPDATED` — статус невыкупа или возврата изменен.
     * * `GOODS_FEEDBACK_CREATED` — создан новый отзыв о товаре.
     * * `GOODS_FEEDBACK_COMMENT_CREATED` — создан новый комментарий к отзыву о товаре.
     * * `CHAT_CREATED` — создан новый чат с покупателем.
     * * `CHAT_MESSAGE_SENT` — добавлено новое сообщение в чате.
     * * `CHAT_ARBITRAGE_STARTED` — по обращению покупателя начался спор.
     * * `CHAT_ARBITRAGE_FINISHED` — спор завершен.
     */
    notificationType: NotificationType
    /**
     * Идентификатор заказа.
     * @format int64
     */
    orderId: number
    /**
     * Идентификатор магазина.
     * @format int64
     */
    campaignId: number
    /** Список товаров в заказе. */
    items: NotificationOrderItemDTO[]
    /**
     * Дата и время отмены заказа.
     *
     * Формат даты: ISO 8601 со смещением относительно UTC. Например, `2017-11-21T00:00:00.213Z`.
     * @format date-time
     */
    cancelledAt: string
}
// ...
export enum NotificationType {
    PING = 'PING',
    ORDER_CREATED = 'ORDER_CREATED',
    ORDER_CANCELLED = 'ORDER_CANCELLED',
    ORDER_STATUS_UPDATED = 'ORDER_STATUS_UPDATED',
    ORDER_RETURN_CREATED = 'ORDER_RETURN_CREATED',
    ORDER_CANCELLATION_REQUEST = 'ORDER_CANCELLATION_REQUEST',
    ORDER_RETURN_STATUS_UPDATED = 'ORDER_RETURN_STATUS_UPDATED',
    GOODS_FEEDBACK_CREATED = 'GOODS_FEEDBACK_CREATED',
    GOODS_FEEDBACK_COMMENT_CREATED = 'GOODS_FEEDBACK_COMMENT_CREATED',
    CHAT_CREATED = 'CHAT_CREATED',
    CHAT_MESSAGE_SENT = 'CHAT_MESSAGE_SENT',
    CHAT_ARBITRAGE_STARTED = 'CHAT_ARBITRAGE_STARTED',
    CHAT_ARBITRAGE_FINISHED = 'CHAT_ARBITRAGE_FINISHED'
}
// ...
export interface NotificationOrderItemDTO {
    /**
     * Ваш SKU — идентификатор товара в вашей системе.
     *
     * Правила использования SKU:
     *
     * * У каждого товара SKU должен быть свой.
     *
     * * Уже заданный SKU нельзя освободить и использовать заново для другого товара. Каждый товар должен получать новый идентификатор, до того никогда не использовавшийся в вашем каталоге.
     *
     * SKU товара можно изменить в кабинете продавца на Маркете. О том, как это сделать, читайте [в Справке Маркета для продавцов](https://yandex.ru/support2/marketplace/ru/assortment/operations/edit-sku).
     *
     * [Что такое SKU и как его назначать](https://yandex.ru/support/marketplace/assortment/add/index.html#fields)
     * @minLength 1
     * @maxLength 255
     * @pattern ^(?=.*\S.*)[^\x00-\x08\x0A-\x1f\x7f]{1,255}$
     */
    offerId: string
    /** Количество товара. */
    count: number
}
// ...
export enum OrderStatusType {
    PLACING = 'PLACING',
    RESERVED = 'RESERVED',
    UNPAID = 'UNPAID',
    PROCESSING = 'PROCESSING',
    DELIVERY = 'DELIVERY',
    PICKUP = 'PICKUP',
    DELIVERED = 'DELIVERED',
    CANCELLED = 'CANCELLED',
    PENDING = 'PENDING',
    PARTIALLY_RETURNED = 'PARTIALLY_RETURNED',
    RETURNED = 'RETURNED',
    UNKNOWN = 'UNKNOWN'
}